Rename from "areas" to "geography"
