+++ src/app/beta/orgs/[orgId]/areaassignments/[areaAssId]/areasgraph/route.ts
@@ -1,7 +1,7 @@
 import mongoose from 'mongoose';
 import { NextRequest, NextResponse } from 'next/server';
 
-import { AreaModel } from 'features/areas/models';
+import { AreaModel } from 'features/geography/models';
 import {
   AreaAssignmentModel,
   PlaceModel,
@@ -18,7 +18,7 @@ import getAreaData from 'features/areaAssignments/utils/getAreaData';
 import isPointInsidePolygon from 'features/areaAssignments/utils/isPointInsidePolygon';
 import asOrgAuthorized from 'utils/api/asOrgAuthorized';
 import { ZetkinPerson } from 'utils/types/zetkin';
-import { ZetkinArea } from 'features/areas/types';
+import { ZetkinArea } from 'features/geography/types';
 
 type RouteMeta = {
   params: {
