+++ src/features/breadcrumbs/l10n/messageIds.ts
@@ -5,7 +5,6 @@ export default makeMessages('feat.breadcrumbs', {
     activities: m('Activities'),
     archive: m('Archive'),
     areaassignments: m('Area assignments'),
-    areas: m('Areas'),
     assignees: m('Assignees'),
     calendar: m('Calendar'),
     callassignments: m('Call assignments'),
@@ -18,6 +17,7 @@ export default makeMessages('feat.breadcrumbs', {
     emails: m('Emails'),
     events: m('Events'),
     folders: m('Lists'),
+    geography: m('Geography'),
     incoming: m('Incoming'),
     insights: m('Insights'),
     instances: m('Instances'),
