+++ src/features/geography/components/AreaOverlay/TagsSection.tsx
@@ -1,9 +1,9 @@
 import { FC, useState } from 'react';
 import { Box, Typography } from '@mui/material';
 
-import useAreaTagging from 'features/areas/hooks/useAreaTagging';
-import useAreaTags from 'features/areas/hooks/useAreaTags';
-import { ZetkinArea } from 'features/areas/types';
+import useAreaTagging from 'features/geography/hooks/useAreaTagging';
+import useAreaTags from 'features/geography/hooks/useAreaTags';
+import { ZetkinArea } from 'features/geography/types';
 import TagManager from 'features/tags/components/TagManager';
 import GroupToggle from 'features/tags/components/TagManager/components/GroupToggle';
 import ZUIFuture from 'zui/ZUIFuture';
