+++ src/pages/organize/[orgId]/projects/[campId]/areaassignments/[areaAssId]/map.tsx
@@ -6,7 +6,7 @@ import { GetServerSideProps } from 'next';
 import { scaffold } from 'utils/next';
 import { PageWithLayout } from 'utils/types';
 import AreaAssignmentLayout from 'features/areaAssignments/layouts/AreaAssignmentLayout';
-import useAreas from 'features/areas/hooks/useAreas';
+import useAreas from 'features/geography/hooks/useAreas';
 import useServerSide from 'core/useServerSide';
 import useAreaAssignmentSessions from 'features/areaAssignments/hooks/useAreaAssignmentSessions';
 import useCreateAreaAssignmentSession from 'features/areaAssignments/hooks/useCreateAreaAssigneeSession';
@@ -15,7 +15,7 @@ import usePlaces from 'features/areaAssignments/hooks/usePlaces';
 import useAssignmentAreaStats from 'features/areaAssignments/hooks/useAssignmentAreaStats';
 import ZUIFutures from 'zui/ZUIFutures';
 import useAreaAssignment from 'features/areaAssignments/hooks/useAreaAssignment';
-import AreaFilterProvider from 'features/areas/components/AreaFilters/AreaFilterContext';
+import AreaFilterProvider from 'features/geography/components/AreaFilters/AreaFilterContext';
 import AssigneeFilterProvider from 'features/areaAssignments/components/OrganizerMapFilters/AssigneeFilterContext';
 
 const OrganizerMap = dynamic(
